Created Category Activity and KMS Activity
Created the category activity that when a faction has a lot of ships, the faction activity directs to this activity.  This category activity redirects to the list of ships of ship type the user selected in faction that was selected.  Added the whole kms faction.
Tried to create separate files so everything is not jam packed into one file.